SoloLeveling_AI/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
discord.py==2.2.2
Django==4.1.7
djangorestframework==3.14.0
filelock==3.9.0
flatbuffers==23.3.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.12.1
idna==3.4
jax==0.4.5
keras==2.12.0rc1
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0rc0
tensorflow-estimator==2.12.0rc0
tensorflow-intel==2.12.0rc0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tokenizers==0.13.2
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
pip install googletrans==3.1.0a0